Album Hierarchy and Album textile Database
Added a package called AlbumHierarchy with Album and Track files that
both extend the same abstract class.

Add a package called Utilities that contains AlbumFileReadWrite, which
has methods for reading and writing albums from a textfile database
(YOU WILL NEED TO CHANGE THE LOCATION OF THE TEXTFILE IN UTILITIES TO A
PATH THAT EXISTS ON YOUR COMPUTER until this problem is patched). This
should be replaced with a SQLight Database that is communicated with
JSON.

The CreateNewAlbum page takes user input and writes it the textfile
database.